Move TagContext serialization classes into io.opencensus.tags.propagation.
Closes #641.
