+++ core/src/test/java/io/opencensus/tags/NoopTagsTest.java
@@ -22,6 +22,7 @@ import com.google.common.collect.Lists;
 import io.opencensus.tags.Tag.TagString;
 import io.opencensus.tags.TagKey.TagKeyString;
 import io.opencensus.tags.TagValue.TagValueString;
+import io.opencensus.tags.propagation.TagContextParseException;
 import java.util.Arrays;
 import java.util.Iterator;
 import org.junit.Test;
