+++ core_impl/src/test/java/io/opencensus/implcore/tags/propagation/TagContextDeserializationTest.java
@@ -14,17 +14,18 @@
  * limitations under the License.
  */
 
-package io.opencensus.implcore.tags;
+package io.opencensus.implcore.tags.propagation;
 
 import static com.google.common.truth.Truth.assertThat;
 
 import io.opencensus.implcore.internal.VarInt;
+import io.opencensus.implcore.tags.TaggerImpl;
 import io.opencensus.tags.TagContext;
-import io.opencensus.tags.TagContextBinarySerializer;
-import io.opencensus.tags.TagContextParseException;
 import io.opencensus.tags.TagKey.TagKeyString;
 import io.opencensus.tags.TagValue.TagValueString;
 import io.opencensus.tags.Tagger;
+import io.opencensus.tags.propagation.TagContextBinarySerializer;
+import io.opencensus.tags.propagation.TagContextParseException;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import org.junit.Test;
