+++ core_impl_android/src/test/java/io/opencensus/impllite/tags/TagsTest.java
@@ -18,8 +18,8 @@ package io.opencensus.impllite.tags;
 
 import static com.google.common.truth.Truth.assertThat;
 
-import io.opencensus.implcore.tags.TagPropagationComponentImpl;
 import io.opencensus.implcore.tags.TaggerImpl;
+import io.opencensus.implcore.tags.propagation.TagPropagationComponentImpl;
 import io.opencensus.tags.Tags;
 import io.opencensus.tags.TagsComponent;
 import org.junit.Test;
