+++ core_impl_java/src/test/java/io/opencensus/impl/tags/TagsTest.java
@@ -18,8 +18,8 @@ package io.opencensus.impl.tags;
 
 import static com.google.common.truth.Truth.assertThat;
 
-import io.opencensus.implcore.tags.TagPropagationComponentImpl;
 import io.opencensus.implcore.tags.TaggerImpl;
+import io.opencensus.implcore.tags.propagation.TagPropagationComponentImpl;
 import io.opencensus.tags.Tags;
 import io.opencensus.tags.TagsComponent;
 import org.junit.Test;
